added Restic test
