+++ tests/e2e/common.go
@@ -2,20 +2,13 @@ package e2e
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
-	"flag"
-	"fmt"
-	"io/ioutil"
-	"time"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/runtime/serializer/yaml"
-	"k8s.io/apimachinery/pkg/util/wait"
 
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/dynamic"
@@ -98,6 +91,10 @@ func createVeleroClient(client dynamic.Interface) (dynamic.ResourceInterface, er
 	return namespaceResClient, nil
 }
 
+func getKubeConfig() *rest.Config {
+	return config.GetConfigOrDie()
+}
+
 func createDefaultVeleroCR(res *unstructured.Unstructured, client dynamic.Interface) (*unstructured.Unstructured, error) {
 	veleroClient, err := createVeleroClient(client)
 	if err != nil {
@@ -118,203 +115,4 @@ func deleteVeleroCR(client dynamic.Interface) error {
 		return err
 	}
 	return veleroClient.Delete(context.Background(), "example-velero", metav1.DeleteOptions{})
-}
-
-func getKubeConfig() *rest.Config {
-	return config.GetConfigOrDie()
-}
-
-func isVeleroPodRunning() wait.ConditionFunc {
-	return func() (bool, error) {
-		kubeConf := getKubeConfig()
-
-		// create client for pod
-		clientset, err := kubernetes.NewForConfig(kubeConf)
-		if err != nil {
-			panic(err)
-		}
-		// select Velero pod with this label
-		veleroOptions := metav1.ListOptions{
-			LabelSelector: "component=velero",
-		}
-		// get pods in the oadp-operator-e2e namespace
-		podList, err := clientset.CoreV1().Pods("oadp-operator").List(context.TODO(), veleroOptions)
-		if err != nil {
-			panic(err)
-		}
-		// get pod name and status with specified label selector
-		var status string
-		for _, podInfo := range (*podList).Items {
-			status = string(podInfo.Status.Phase)
-		}
-		if status != "Running" {
-			fmt.Println("Checking pod status...")
-		}
-		if status == "Running" {
-			fmt.Println("Velero pod is running")
-			return true, nil
-		}
-		return false, err
-	}
-}
-
-func waitForVeleroPodRunning() error {
-	// poll pod every 5 secs for 2 mins until it's running or timeout occurs
-	return wait.PollImmediate(time.Second*5, time.Minute*2, isVeleroPodRunning())
-}
-
-func areResticPodsRunning() wait.ConditionFunc {
-	return func() (bool, error) {
-		kubeConf := getKubeConfig()
-		// create client for daemonset
-		client, err := kubernetes.NewForConfig(kubeConf)
-		if err != nil {
-			return false, err
-		}
-		resticOptions := metav1.ListOptions{
-			FieldSelector: "metadata.name=restic",
-		}
-		// get daemonset in oadp-operator-e2e ns with specified field selector
-		resticDaemeonSet, err := client.AppsV1().DaemonSets("oadp-operator").List(context.TODO(), resticOptions)
-		if err != nil {
-			return false, err
-		}
-		var numScheduled int32
-		var numDesired int32
-
-		for _, daemonSetInfo := range (*resticDaemeonSet).Items {
-			numScheduled = daemonSetInfo.Status.CurrentNumberScheduled
-			numDesired = daemonSetInfo.Status.DesiredNumberScheduled
-		}
-		// if numScheduled == numDesired, then all restic pods are running
-		if numScheduled != 0 && numDesired != 0 {
-			if numScheduled != numDesired {
-				fmt.Println("Checking pod status...")
-			}
-			if numScheduled == numDesired {
-				fmt.Println("All restic pods are running")
-				return true, nil
-			}
-		}
-		return false, err
-	}
-}
-
-func waitForResticPods() error {
-	// poll pod every 5 secs for 2 mins until it's running or timeout occurs
-	return wait.PollImmediate(time.Second*5, time.Minute*2, areResticPodsRunning())
-}
-
-func isVeleroCRFailed() wait.ConditionFunc {
-	kubeConfig := getKubeConfig()
-
-	// create dynamic client for CR
-	client, err := dynamic.NewForConfig(kubeConfig)
-	if err != nil {
-		return nil
-	}
-	veleroClient, err := createVeleroClient(client)
-	if err != nil {
-		return nil
-	}
-	return func() (bool, error) {
-		// Get velero CR in cluster
-		veleroResource, err := veleroClient.Get(context.Background(), "example-velero", metav1.GetOptions{})
-		if err != nil {
-			return false, err
-		}
-		// Read status subresource from cluster
-		veleroStatus := ansibleOperatorStatus{}
-		statusObj := veleroResource.Object["status"]
-		// Convert status subresource interface to typed structure
-		statusBytes, err := json.Marshal(statusObj)
-		if err != nil {
-			return false, err
-		}
-		err = json.Unmarshal(statusBytes, &veleroStatus)
-		if err != nil {
-			return false, err
-		}
-		conditions := veleroStatus.Conditions
-		var message string
-
-		for _, condition := range conditions {
-			message = condition.Message
-			if condition.Type != "Failure" {
-				fmt.Println("Checking Velero status...")
-			}
-			if condition.Type == "Failure" {
-				fmt.Printf("Velero install failure: %s\n", message)
-				return true, nil
-			}
-		}
-		return false, nil
-	}
-}
-
-func waitForFailedVeleroCR() error {
-	return wait.PollImmediate(time.Second*5, time.Minute*2, isVeleroCRFailed())
-}
-
-type ansibleOperatorStatus struct {
-	Conditions []condition `json:"conditions"`
-}
-type condition struct {
-	AnsibleResult ansibleResult `json:"ansibleResult,omitempty"`
-	//LastTransitionTime time.Time     `json:"lastTransitionTime"`
-	Message string `json:"message"`
-	Reason  string `json:"reason"`
-	Status  string `json"status"`
-	Type    string `json:"type"`
-}
-
-type ansibleResult struct {
-	Changed int `json:"changed"`
-	//Completion time.Time `json:"completion"`
-	Failures int `json:"failures"`
-	Ok       int `json:"ok"`
-	Skipped  int `json:"skipped"`
-}
-
-func getCredsData() []byte {
-	// pass in aws credentials by cli flag
-	// from cli:  -cloud=<"filepath">
-	// go run main.go -cloud="/Users/emilymcmullan/.aws/credentials"
-	cloud := flag.String("cloud", "", "file path for aws credentials")
-	flag.Parse()
-
-	// save passed in cred file as []byte
-	credsFile, err := ioutil.ReadFile(*cloud)
-	if err != nil {
-		panic(err)
-	}
-	return credsFile
-}
-
-func createSecret(data []byte) error {
-	config := getKubeConfig()
-	clientset, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		return err
-	}
-	sec := corev1.Secret{
-		ObjectMeta: v1.ObjectMeta{
-			Name:      "cloud-credentials",
-			Namespace: "oadp-operator",
-		},
-		TypeMeta: v1.TypeMeta{
-			Kind:       "Secret",
-			APIVersion: v1.SchemeGroupVersion.String(),
-		},
-		Data: map[string][]byte{
-			"cloud": data,
-		},
-		Type: corev1.SecretTypeOpaque,
-	}
-	_, errors := clientset.CoreV1().Secrets("oadp-operator").Create(context.TODO(), &sec, v1.CreateOptions{})
-	if apierrors.IsAlreadyExists(errors) {
-		fmt.Println("Secret already exists in this namespace")
-		return nil
-	}
-	return err
 }
